+++ packages/webview/src/env.d.ts
@@ -1,9 +1,0 @@
-interface vscode {
-  postMessage(message: any): void;
-}
-
-declare global {
-  function acquireVsCodeApi(): vscode;
-}
-
-export {};