+++ src/webSocket.ts
@@ -1,16 +1,17 @@
-import io from 'socket.io-client';
 import * as vscode from 'vscode';
 import { WombatOutputChannel } from './wombatOutputChannel';
-import { SidebarProvider } from './sidebarProvider';
+import { TreeViewProvider } from './treeViewProvider';
 import { API } from './api';
 import { ConnectionService } from './connectionService';
+import io from 'socket.io-client';
+import { AddressService } from './addressService';
 
 export class WebSocket {
-    private sidebar;
+    private readonly treeView: TreeViewProvider;
     private socket?: SocketIOClient.Socket = undefined;
 
     constructor(context: vscode.ExtensionContext, private connectionService: ConnectionService) {
-        this.sidebar = SidebarProvider.getInstance(context);
+        this.treeView = TreeViewProvider.getInstance(context);
     }
 
     /**
@@ -23,7 +24,7 @@ export class WebSocket {
             this.socket = undefined;
         }
 
-        this.socket = io(`ws://${API.address}/runner`, {
+        this.socket = io(`ws://${AddressService.address}/runner`, {
             reconnection: true,
             reconnectionAttempts: Infinity,
             reconnectionDelay: 50,
@@ -31,6 +32,8 @@ export class WebSocket {
             timeout: 200,
             autoConnect: true,
         } as SocketIOClient.ConnectOpts);
+
+        this.socket.connect();
 
         this.socket.on('stdout', (line: string) => {
             WombatOutputChannel.print(line);
@@ -44,7 +47,7 @@ export class WebSocket {
         this.socket.on('connect', () => {
             vscode.window.showInformationMessage('Connected to Wombat');
             this.connectionService.setConnectionStatus(true);
-            this.sidebar.refresh();
+            this.treeView.refresh();
         });
     }
 }